Replace fragile string prefix checks with `strhasprefix()`
This commit introduces two new functions to `string-utils.h`:

*   `strhasprefix(cstr, prefix)`: Returns true if the c-string `cstr`
    is prefixed by the c-string `prefix`.
*   `strcasehasprefix(cstr, prefix)`: Case insensitive version of
    `strhasprefix()`.

Previously, we were using `strncmp()` for this sort of thing. However,
this is dangerous because when copying and pasting code it is easy to
forget to update the prefix length argument. Indeed, this has happened
several times. Using `strhasprefix()` will eliminate such problems.

https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=3448#c3